feat(nextday): 分钟 HGB 多年份滚动稳健性验证
- 4 个独立样本外 test 年（2022-2025）全部为正：Rank IC 0.022~0.035
- run_minute_baseline.py 按年流式读取（_build_samples_by_year），解决多年数据 OOM，峰值内存从 20GB+ 降至约 9GB
- minute_baseline.py 分钟特征 float64 转 float32，内存减半、指标不变
- 新增滚动配置与文档第 9 节

+++ src/ticknet/nextday/minute_baseline.py
@@ -180,7 +180,7 @@ def _stream_l2_modality(
         minutes = minutes[mask]
         feature_matrix = np.column_stack(
             [
-                table[column].to_numpy(zero_copy_only=False).astype(np.float64, copy=False)[mask]
+                table[column].to_numpy(zero_copy_only=False).astype(np.float32, copy=False)[mask]
                 for column in features
             ]
         )
@@ -284,7 +284,7 @@ def read_tushare_minute_rows(
             minute_of_day = np.mod(epoch_minutes, 24 * 60)
             feature_matrix = np.column_stack(
                 [
-                    table[column].to_numpy(zero_copy_only=False).astype(np.float64, copy=False)
+                    table[column].to_numpy(zero_copy_only=False).astype(np.float32, copy=False)
                     for column in TUSHARE_FEATURE_COLUMNS
                 ]
             )
